Remove SelectedContainerDefinition and SelectedContainer from Selectors
